+++ backend/storage/bolt/users.go
@@ -60,16 +60,25 @@ func (st usersBackend) Gets() ([]*users.User, error) {
 
 func (st usersBackend) Update(user *users.User, fields ...string) error {
 	if len(fields) == 0 {
-		return st.Save(user)
+		return nil
 	}
 
-	val := reflect.ValueOf(user).Elem()
+	userValue := reflect.ValueOf(user).Elem() // Get reflect.Value of user struct
+	var target reflect.Value
+
+	if user.Perm.Admin {
+		target = userValue // Admins can update all fields
+	} else {
+		target = userValue.FieldByName("NonAdminEditable") // Non-admins can only update NonAdminEditable fields
+		if !target.IsValid() {
+			return fmt.Errorf("NonAdminEditable struct not found")
+		}
+	}
 
 	for _, field := range fields {
-		// Capitalize the first letter (you can adjust this based on your field naming convention)
 		correctedField := utils.CapitalizeFirst(field)
 
-		userField := val.FieldByName(correctedField)
+		userField := target.FieldByName(correctedField)
 		if !userField.IsValid() {
 			return fmt.Errorf("invalid field: %s", field)
 		}
@@ -77,11 +86,12 @@ func (st usersBackend) Update(user *users.User, fields ...string) error {
 			return fmt.Errorf("cannot update unexported field: %s", field)
 		}
 
-		val := userField.Interface()
-		if err := st.db.UpdateField(user, correctedField, val); err != nil {
-			return fmt.Errorf("failed to update user field: %s, error: %v", correctedField, err.Error())
+		fieldValue := userField.Interface()
+		if err := st.db.UpdateField(user, correctedField, fieldValue); err != nil {
+			return fmt.Errorf("failed to update user field: %s, error: %v", correctedField, err)
 		}
 	}
+
 	return nil
 }